Replace CLI with GUI
The CLI is not intuitive for people unfamiliar with CLI

Let's implement a GUI using JavaFX and store them in mirai.ui package,
which will replace the original Ui class in the CLI version.
Refactor Command functional interface to return a message instead of a
flag to end or continue the conversation. Update methods in Parser
implementing this Command interface.

This is because the GUI does not need a loop to maintain the
conversation, unlike the CLI that contains a while loop under the hood
and needs a boolean flag to continue or terminate the loop. Instead
of the UI class printing the message to the terminal, a Command
directly returns the String message to feed into the DialogBox.