Improve GUI and flexfind algorithm
The GUI is generic.

Choosing another colour theme for the GUI will make it less generic
and show that there is a certain theme associated with the chatbot.

Let's add styling to the JavaFX components.

CSS is used for styling because it concentrates all styling into
one file and style changes will be easier. Blue theme is used
because the imagined avatar of Mirai is a girl with blue hair and
eyes.

Next, the flexfind algorithm has a flaw. For each keyword, it
performs edit distance calculation and return the distance from
the closest word. For n keywords, the closeness is the sum of
n edit distances. However, when flexfinding with 'r', 'wake up'
ranks higher than 'read book' because 'r' is closer to 'up' than
'read'.

The improved algorithm must also account for inclusion, rather than
just edit distance to the closest word.

Let's add another logic layer to the algorithm, where if a description
word contains a key word, edit distance is straight up 0.